Moved test files alongside their files
