+++ site/components/layout/PhaseBanner.test.tsx
@@ -1,11 +1,11 @@
 import * as React from "react";
 import renderer from "react-test-renderer";
-import Layout from "../../components/layout/Layout";
+import PhaseBanner from "./PhaseBanner";
 import { describe, it, expect } from "vitest";
 
-describe("Layout", () => {
+describe("PhaseBanner", () => {
     it("should render correctly", () => {
-        const tree = renderer.create(<Layout title="title" description="description" />).toJSON();
+        const tree = renderer.create(<PhaseBanner />).toJSON();
         expect(tree).toMatchSnapshot();
     });
 });
