+++ site/pages/index.test.tsx
@@ -1,6 +1,6 @@
 import * as React from "react";
 import renderer from "react-test-renderer";
-import Index from "../pages/index";
+import Index from "./index";
 import { describe, it, expect } from "vitest";
 
 describe("pages", () => {
