+++ site/styles/globals.scss
@@ -3,7 +3,3 @@
 @tailwind base;
 @tailwind components;
 @tailwind utilities;
-
-.govuk-footer__copyright-logo {
-    background-image: url('../public/assets/images/govuk-crest-2x.png') !important;
-}
